merge -r11468 from branch 2.3 to trunk
New: 2 Perl TUJs to check tFileInputPositional with basic settings and with
the "trim all columns" option. tFIP_basics logically fail because the unpack
function automatically trim (which is the opposite of what the GUI let the
user think), this will be fixed on trunk when using substr function instead
of unpack.


git-svn-id: http://talendforge.org/svn/tos/trunk@11469 f6f1c999-d317-4740-80b0-e6d1abc6f99e